+++ src/main/java/com/devbliss/doctest/httpwithoutredirect/DeleteWithoutRedirectImpl.java
@@ -1,5 +1,8 @@
 package com.devbliss.doctest.httpwithoutredirect;
 
+import static com.devbliss.doctest.httpwithoutredirect.HttpConstants.ENCODING;
+import static com.devbliss.doctest.httpwithoutredirect.HttpConstants.HANDLE_REDIRECTS;
+
 import java.io.IOException;
 import java.net.URI;
 
@@ -15,8 +18,14 @@ import com.google.inject.Inject;
 import de.devbliss.apitester.factory.DeleteFactory;
 import de.devbliss.apitester.factory.HttpDeleteWithBody;
 
+/**
+ * Implements a DELETE HTTP request which does not handle any redirect.
+ * 
+ * @author bmary
+ * 
+ */
 public class DeleteWithoutRedirectImpl implements DeleteFactory {
-    private static final String ENCODING = "UTF-8";
+
     private final Gson gson;
 
     @Inject
@@ -27,7 +36,7 @@ public class DeleteWithoutRedirectImpl implements DeleteFactory {
     public HttpDeleteWithBody createDeleteRequest(URI uri, Object payload) throws IOException {
         HttpDeleteWithBody httpDelete = new HttpDeleteWithBody(uri);
         HttpParams params = new BasicHttpParams();
-        params.setParameter("http.protocol.handle-redirects", false);
+        params.setParameter(HANDLE_REDIRECTS, false);
         httpDelete.setParams(params);
 
         if (payload != null) {
@@ -43,7 +52,7 @@ public class DeleteWithoutRedirectImpl implements DeleteFactory {
     public HttpDelete createDeleteRequest(URI uri) throws IOException {
         HttpDelete httpDelete = new HttpDelete(uri);
         HttpParams params = new BasicHttpParams();
-        params.setParameter("http.protocol.handle-redirects", false);
+        params.setParameter(HANDLE_REDIRECTS, false);
         httpDelete.setParams(params);
         return httpDelete;
     }