+++ src/main/java/com/devbliss/doctest/httpwithoutredirect/HttpConstants.java
@@ -1,0 +1,13 @@
+package com.devbliss.doctest.httpwithoutredirect;
+
+/**
+ * Defines some constants which are used in several classes of the project.
+ * 
+ * @author bmary
+ * 
+ */
+class HttpConstants {
+
+    static final String ENCODING = "UTF-8";
+    static final String HANDLE_REDIRECTS = "http.protocol.handle-redirects";
+}